refactor: adds strategy pattern to contentservice
